Rename exceptions modules to avoid stdlib module name collision.
I forgot about the exceptions module in the standard library, so
I had a module naming collision. I don't know if it actually is
an issue as my exceptions was in my own package, but I think
renaming is still prudent as it will avoid unforseen consequences
down the line.