Change random_account return signature from tuple to dict.
Breaks backwards compatibility.

Alters the return signature of the main random_account function to
return a dict of the account information instead of the original
tuple. Update function docstring to reflect the changes. Exposes
constants that are used to define the keys of the returned dict:
  - PROVIDER = 'provider'
  - USERNAME = 'username'
  - PASSWORD = 'password'
  - EMAIL = 'email'

Add the provider string to the return dict as APIs require an
indication of which type of account is being used to log in,
currently 'ptc' or 'google'.

Design decisions based of of presumptious use of tejado/pgoapi.
 * Provider string 'ptc' one of two accepted options for this API
 * Default return dict key strings are set in such a way that the
     returned dict can be unpacked directly into the api login
     method.
     An illustration:
       api = pgoapi.PGoApi()
       api.login(**ptcaccount.random_account())

Closes issue #5